add project variable NETWORK_THEME
to be able to distinguish between transport properties from the common
transport elements schema if they are used for more than one network
theme. The value of the variable can be empty if the target datasets are
stored separately. It should be Air, Cable, Railway, Road or Water if
the distinction is needed.