feat(tn-ro): add project variable to limit number of network references
Setting the new project variable `LIMIT_NETWORK_REFERENCES` reduces the
number of object references in the network objects to the number set
in the variable and the `TransportNetwork` object will contain no references
at all.

If the variable is not set, the number of references is not limited
(original behaviour).