fix: adapt documentation according to comment experts
The values documented for LN_StrassenverkehrUndWegeverkehr.Funktion have
not been correct in the mapping table. Some were missing and some had
wrong codes. The current codes can be found in the screenshot that is
added as well and was communicated by the experts here
https://wetransform.atlassian.net/browse/WGS-901?focusedCommentId=51160
There was no need to adapt the alignment as all
LN_StrassenverkehrUndWegeverkehr objects get the same land use value
assigned.

SVC-867 - WGS-901