fix(tn-w): fix ConditionOfWaterFacility associations
Replaces the association of `ConditionOfWaterFacility` objects to
`WaterwayLink` objects with associations to `ConditionOfFacility`
objects because an association to `ConditionOfWaterFacility` is not
allowed by the INSPIRE data model (see sec. 5.7.2.1.4. of the Data
Specification on Transport Networks - Technical Guidelines).

Also, an association of `PortArea` objects to `ConditionOfWaterFacility`
is added for which such an association is allowed.

SVC-1192