refactor: replace line break symbols in groovy scripts in base alignment
Opened and saved alignment with hs 5.1.0 snapshot to make sure that the
line breaks use the ones implemented in hs 5.0. Doing this in a separate
commit to allow easier reviewing actual content changes.
The HY-P transformation is running into transformation errors caused by
the custom functions and I want to make sure it is not caused by the
fact that the line breaks are not read correctly.

SVC-1480