fix: make sure name instances are only created if value is existing
With the old implementation, the groovy script returned a
GeographicalName instance independent from if there was one (or more)
source values.
Also, the potential occurrence of multiple regionalsprache or zweitname
instances has not been reflected.
For GeographicalName instances resulting from zweitname, the language is
set to German, for those resulting from regionalsprache, it is empty
with a nilReason (assuming that regionalsprache is not German).

SVC-1480