fix: remove non-existing bauwerksfunktion code 1800 from alignment
In the original alignment, a CQL filter on AX_BauwerkImGewaesserbereich.bauwerksfunktion=1800 was contained. Such a code does not exist for that attribute according
to the Objektartenkatalog. Replaced it with the filter for 2012 (which is existing) and which should be used according to the original mapping table https://github.com/wetransform/adv-inspire-alignments/blob/master/annex-1/mappings6.0/Hydrography/Vorgaben/Basis-DLM/20160407_HY_MappingTable_Network.xls.

SVC-1471