fix: move mappings to filtered attributes instead of filtered objects
In the alignment, some attribute mappings had been defined in types using an
instance context, but with no mapping defined on the types themselves.
That lead to these mappings not having any effect. They are now changed in a
way that they are based on filtered attributes in the unfiltered type.
This change will also need to be done for the 6.0 alignments.

SVC-1478