fix: adapt ad-hk Gradle transformation to new source file format
Format version 5.0 of the Hauskoordinaten source file no longer requires a
"Schluessel" file, so the correspondig configuration in the Gradle build is
removed. The new format has a header line, so the `skipFirst` default is
changed accordingly. The `skip` parameter is also changed to a number as
the corresponding parameter of the hale»studio CSV reader has changed.

SVC-1463