feat: use codelists instead of enumerations
According to [this schema release](https://github.com/INSPIRE-MIF/application-schemas/releases/tag/2024.1),
the enumerations used in the INSPIRE schemas are replaced by codelists.
The alignments using enumerations so far are adapted accordingly including
import of the codelists into the alignment, assigning them to the respective
elements and assigning the matching value.

SVC-1946