feat: add solution for TN-A and TN-C alignments
The INSPIRE schema update to version 5.0 had impacts on the base alignments
as well as the TN base alignments.
As there was no new versions for the TN-A and TN-C schemas, the JRC decided
to also keep the TN common schema version 4.0 for it.
That means that the migrated base alignments and TN base alignments now using
version 5.0 cannot longer be used for TN-A and TN-C.
Therefore, new 4.0 versions of these alignments were added and used in TN-A
and TN-C.

SVC-1946